osrs_rl_gui/requirements.txt
PySide6
PyYAML
cryptography
